+++ osu.Framework/Testing/Drawables/TestCaseButton.cs
@@ -9,7 +9,6 @@ using System.Reflection;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.UserInterface;
 using osuTK.Graphics;
 using Container = osu.Framework.Graphics.Containers.Container;
