+++ osu.Framework/Testing/Drawables/TestCaseSubButton.cs
@@ -6,7 +6,6 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.UserInterface;
 
 namespace osu.Framework.Testing.Drawables
 {
